Integrador con excepciones y test de componente

+++ clinicaOdontologica/src/main/java/com/example/clinicaOdontologica/service/impl/OdontologoServiceImpl.java
@@ -1,5 +1,7 @@
 package com.example.clinicaOdontologica.service.impl;
 
+import com.example.clinicaOdontologica.exceptions.ServiceBadRequestException;
+import com.example.clinicaOdontologica.exceptions.ServiceNotFoundException;
 import com.example.clinicaOdontologica.model.OdontologoDTO;
 import com.example.clinicaOdontologica.persistence.entities.Odontologo;
 import com.example.clinicaOdontologica.persistence.repositories.OdontologoRepository;
@@ -32,22 +34,26 @@ public class OdontologoServiceImpl implements OdontologoService {
     }
 
     @Override
-    public OdontologoDTO buscar(Integer id) {
+    public OdontologoDTO buscar(Integer id) throws ServiceBadRequestException, ServiceNotFoundException{
         log.debug("Iniciando método buscar odontólogo");
-        OdontologoDTO odontologoBuscado = null;
 
-        try{
-            odontologoBuscado = new OdontologoDTO(repository.getById(id));
-        }catch (Exception e){
-            log.error(e.getMessage());
+        if(id == null || id<=0){
+            throw new ServiceBadRequestException("El id del odontólogo debe ser un numero entero mayor a cero");
         }
 
+        Odontologo odontologoBuscado = repository.findById(id).orElse(null);
+
+        if(odontologoBuscado == null){
+            throw new ServiceNotFoundException("No se encontró el odontólogo con id "+id);
+        }
+
+        OdontologoDTO odontologoDto = new OdontologoDTO(odontologoBuscado);
         log.debug("Terminó la ejecución del método buscar odontólogo");
-        return odontologoBuscado;
+        return odontologoDto;
     }
 
     @Override
-    public List<OdontologoDTO> buscarTodos() {
+    public List<OdontologoDTO> buscarTodos() throws ServiceNotFoundException{
         log.debug("Buscando todos los odontólogos");
         List<OdontologoDTO>odontologos = new ArrayList<>();
 
@@ -55,24 +61,34 @@ public class OdontologoServiceImpl implements OdontologoService {
             odontologos.add(new OdontologoDTO(odon));
         }
 
+        if(odontologos.size()==0){
+            throw new ServiceNotFoundException("No existen odontólogos.");
+        }
+
         log.debug("Terminó la búsqueda de todos los odontólogos");
         return odontologos;
     }
 
     @Override
-    public OdontologoDTO actualizar(OdontologoDTO odon) {
+    public OdontologoDTO actualizar(OdontologoDTO odon) throws ServiceNotFoundException, ServiceBadRequestException {
         log.debug("Actualizando odontólogo");
-        OdontologoDTO odontologoActualizado = null;
-
-        odontologoActualizado = new OdontologoDTO(repository.save(odon.toEntity()));
+        OdontologoDTO odontologoBuscado = buscar(odon.getId());
+        OdontologoDTO odontologoActualizado = new OdontologoDTO(repository.save(odon.toEntity()));
 
         log.debug("Finalizó la ejecución del método actualizar odontólogo");
         return odontologoActualizado;
     }
 
     @Override
-    public void eliminar(Integer id) {
+    public void eliminar(Integer id) throws ServiceNotFoundException, ServiceBadRequestException {
         log.debug("Eliminando odontólogo");
+
+        if(id <= 0){
+            throw new ServiceBadRequestException("El id del odontólogo debe ser un numero entero mayor a 0");
+        }
+        if(buscar(id) ==null){
+            throw new ServiceNotFoundException("No existe el odontólogo con id "+ id);
+        }
 
         repository.deleteById(id);
